ui-charts: initialize useSyncManchette
